Initial automatic set constant definition
Adds the code needed to define set constants automatically for a datatype
definition. For a type "('a, 'b) tree", it auto-defines two functions of type
"('a, 'b) tree -> 'a -> bool" and "('a, 'b) tree -> 'b -> bool", and registers
them in an "extra" field in the TypeBase.

To do: integrate this properly, call it automatically on datatype definitions,
and figure out how/where to do it for the initial datatypes (pair, unit, etc).